% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{getProfile}
\alias{getProfile}
\title{Read CLIVAR profile data from DB}
\usage{
getProfile(cruise, station)
}
\arguments{
\item{cruise}{A character value with the whpid cruise name}

\item{station}{A numeric with the station number. Retrieve
all stations if not given.}
}
\value{
A data frame with profile data
}
\description{
Read CLIVAR profile data from DB
}
