% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{getProfileSR}
\alias{getProfileSR}
\title{Read CLIVAR profile data from snics_results in DB}
\usage{
getProfileSR(cruise, station)
}
\arguments{
\item{cruise}{A character value with the whpid cruise name}

\item{station}{A numeric with the station number}
}
\value{
A data frame with profile data
}
\description{
Read CLIVAR profile data from snics_results in DB
}

