% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{getStations}
\alias{getStations}
\title{Read stations for a cruise from database}
\usage{
getStations(cruise, hasdata = TRUE)
}
\arguments{
\item{cruise}{A character vector with cruise name (whpid)}

\item{hasdata}{Checks that data is in os table if TRUE}
}
\value{
A data frame with station names
}
\description{
Read stations for a cruise from database
}
