% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{getWheelCLIVAR}
\alias{getWheelCLIVAR}
\title{Read CLIVAR profile data for a wheel from DB}
\usage{
getWheelCLIVAR(wheel)
}
\arguments{
\item{A}{character value with the wheel name}
}
\value{
A data frame with profile data
}
\description{
Read CLIVAR profile data for a wheel from DB
}

