% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotProfileSR}
\alias{plotProfileSR}
\title{Plot profile data from snics_results for a station}
\usage{
plotProfileSR(profile)
}
\arguments{
\item{profile}{A datatable in the format of getProfileSR output}
}
\value{
A ggplot object
}
\description{
Plot profile data from snics_results for a station
}
