% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSection}
\alias{plotSection}
\title{Plot station locations from a section}
\usage{
plotSection(section)
}
\arguments{
\item{section}{A character vector containing a section name.}
}
\value{
A ggplot object
}
\description{
Plot station locations from a section
}
