% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotWheelProfile}
\alias{plotWheelProfile}
\title{Plot CLIVAR data from a wheel}
\usage{
plotWheelProfile(wheel)
}
\arguments{
\item{wheel}{A wheel name in format: "USAMS010101".}
}
\value{
A ggplot object
}
\description{
Plot CLIVAR data from a wheel
}
