% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotWheelProfile}
\alias{plotWheelProfile}
\title{Plot a depth profile from a station}
\usage{
plotWheelProfile(profile)
}
\arguments{
\item{profile}{A datatable in the format of getProfile output}
}
\value{
A ggplot object
}
\description{
Plot a depth profile from a station
}

